Smart tab completion
It took me way more than I thought to come with a version that sort of
works. Mistakes were: not understanding how the stdlib of Lua was
available (strsub is not on Lua 5.1) not RTFM on Lua :sub method, not
understanding how to send raw keystrokes for the C-P key or tab.